+++ src/components/contact-form/contact-form.scss
@@ -62,7 +62,7 @@
 
   .contact-name {
     display: flex;
-    flex-direction: row !important;
+    // flex-direction: row !important;
     gap: 20px;
 
     .first-name,
