+++ src/components/navigation/navigation.scss
@@ -1,5 +1,5 @@
 .navigation-container {
-  position: sticky;
+  position: fixed;
   top: 0;
   display: flex;
   justify-content: center;
