+++ src/index.scss
@@ -25,7 +25,6 @@ body,
 
 html {
   overflow: none;
-
   &::-webkit-scrollbar {
     display: none;
   }
@@ -33,13 +32,13 @@ html {
 
 main {
   flex: 1;
+  margin-top: 70px;
 }
 
 *,
 *::before,
 *::after {
   box-sizing: border-box;
-  // outline: 1px solid aqua;
 }
 
 h1,
@@ -67,6 +66,7 @@ h2 {
   color: #9ea6df;
   margin: 0;
 }
+
 .h2-heading {
   font-size: 6rem;
   line-height: 5rem;
@@ -154,7 +154,7 @@ ul {
   }
 
   .h2-heading {
-    font-size: rem;
+    font-size: 5.3rem;
     line-height: 5rem;
   }
 }
