+++ src/routes/contact/contact-page.scss
@@ -139,6 +139,9 @@
     h2 {
       margin-bottom: 10px;
     }
+    p {
+      margin: 10px 0 30px;
+    }
   }
 
   .contact-page-form-container {
