+++ src/sections/services-section/services-section.scss
@@ -60,6 +60,7 @@
   justify-content: space-between;
   align-items: center;
   aspect-ratio: 1 / 1;
+  // background: #ecedfc;
   background: rgba(210, 218, 245, 0.32);
   border-radius: 20px;
   box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
